requirements.txt
streamlit
openai
httpx
langchain-text-splitters
langchain-community
langchain_chroma
chromadb
numpy
markdown
beautifulsoup4
pandas
matplotlib
requests

